enisa-nlp/requirements.txt
nltk==3.3
elasticsearch==6.3.0
progressbar-latest==2.4
scikit-learn==0.19.1
scipy==1.1.0
numpy==1.14.5
pandas==0.23.1
beautifulsoup4==4.6.0